+++ Main.py
@@ -1,8 +1,5 @@
 import sys
-if sys.version_info[0] < 3:
-    import got
-else:
-    import got3 as got
+import got
 
 def main():
