+++ got/__init__.py
@@ -1,2 +1,2 @@
-import models
-import manager
+from . import models
+from . import manager
