+++ got/manager/TweetManager.py
@@ -1,4 +1,4 @@
-import urllib,urllib2,json,re,datetime,sys,cookielib
+import urllib.request, urllib.parse, urllib.error,urllib.request,urllib.error,urllib.parse,json,re,datetime,sys,http.cookiejar
 from .. import models
 from pyquery import PyQuery
 
@@ -13,10 +13,7 @@ class TweetManager:
 	
 		results = []
 		resultsAux = []
-		cookieJar = cookielib.CookieJar()
-		
-		if hasattr(tweetCriteria, 'username') and (tweetCriteria.username.startswith("\'") or tweetCriteria.username.startswith("\"")) and (tweetCriteria.username.endswith("\'") or tweetCriteria.username.endswith("\"")):
-			tweetCriteria.username = tweetCriteria.username[1:-1]
+		cookieJar = http.cookiejar.CookieJar()
 
 		active = True
 
@@ -35,29 +32,39 @@ class TweetManager:
 				tweetPQ = PyQuery(tweetHTML)
 				tweet = models.Tweet()
 				
-				usernameTweet = tweetPQ("span:first.username.u-dir b").text();
+				usernameTweet = tweetPQ("span.username.js-action-profile-name b").text();
 				txt = re.sub(r"\s+", " ", tweetPQ("p.js-tweet-text").text().replace('# ', '#').replace('@ ', '@'));
 				retweets = int(tweetPQ("span.ProfileTweet-action--retweet span.ProfileTweet-actionCount").attr("data-tweet-stat-count").replace(",", ""));
 				favorites = int(tweetPQ("span.ProfileTweet-action--favorite span.ProfileTweet-actionCount").attr("data-tweet-stat-count").replace(",", ""));
 				dateSec = int(tweetPQ("small.time span.js-short-timestamp").attr("data-time"));
 				id = tweetPQ.attr("data-tweet-id");
 				permalink = tweetPQ.attr("data-permalink-path");
+				user_id = int(tweetPQ("a.js-user-profile-link").attr("data-user-id"))
 				
 				geo = ''
 				geoSpan = tweetPQ('span.Tweet-geo')
 				if len(geoSpan) > 0:
 					geo = geoSpan.attr('title')
-				
+				urls = []
+				for link in tweetPQ("a"):
+					try:
+						urls.append((link.attrib["data-expanded-url"]))
+					except KeyError:
+						pass
 				tweet.id = id
 				tweet.permalink = 'https://twitter.com' + permalink
 				tweet.username = usernameTweet
+				
 				tweet.text = txt
 				tweet.date = datetime.datetime.fromtimestamp(dateSec)
+				tweet.formatted_date = datetime.datetime.fromtimestamp(dateSec).strftime("%a %b %d %X +0000 %Y")
 				tweet.retweets = retweets
 				tweet.favorites = favorites
 				tweet.mentions = " ".join(re.compile('(@\\w*)').findall(tweet.text))
 				tweet.hashtags = " ".join(re.compile('(#\\w*)').findall(tweet.text))
 				tweet.geo = geo
+				tweet.urls = ",".join(urls)
+				tweet.author_id = user_id
 				
 				results.append(tweet)
 				resultsAux.append(tweet)
@@ -78,33 +85,27 @@ class TweetManager:
 	
 	@staticmethod
 	def getJsonReponse(tweetCriteria, refreshCursor, cookieJar, proxy):
-		url = "https://twitter.com/i/search/timeline?f=tweets&q=%s&src=typd&max_position=%s"
+		url = "https://twitter.com/i/search/timeline?f=tweets&q=%s&src=typd&%smax_position=%s"
 		
 		urlGetData = ''
-		
 		if hasattr(tweetCriteria, 'username'):
 			urlGetData += ' from:' + tweetCriteria.username
-		
-		if hasattr(tweetCriteria, 'querySearch'):
-			urlGetData += ' ' + tweetCriteria.querySearch
-		
-		if hasattr(tweetCriteria, 'near'):
-			urlGetData += "&near:" + tweetCriteria.near + " within:" + tweetCriteria.within
-		
+			
 		if hasattr(tweetCriteria, 'since'):
 			urlGetData += ' since:' + tweetCriteria.since
 			
 		if hasattr(tweetCriteria, 'until'):
 			urlGetData += ' until:' + tweetCriteria.until
-		
-
-		if hasattr(tweetCriteria, 'topTweets'):
-			if tweetCriteria.topTweets:
-				url = "https://twitter.com/i/search/timeline?q=%s&src=typd&max_position=%s"
-		
-		
-		
-		url = url % (urllib.quote(urlGetData), refreshCursor)
+			
+		if hasattr(tweetCriteria, 'querySearch'):
+			urlGetData += ' ' + tweetCriteria.querySearch
+			
+		if hasattr(tweetCriteria, 'lang'):
+			urlLang = 'lang=' + tweetCriteria.lang + '&'
+		else:
+			urlLang = ''
+		url = url % (urllib.parse.quote(urlGetData), urlLang, refreshCursor)
+		#print(url)
 
 		headers = [
 			('Host', "twitter.com"),
@@ -117,19 +118,21 @@ class TweetManager:
 		]
 
 		if proxy:
-			opener = urllib2.build_opener(urllib2.ProxyHandler({'http': proxy, 'https': proxy}), urllib2.HTTPCookieProcessor(cookieJar))
+			opener = urllib.request.build_opener(urllib.request.ProxyHandler({'http': proxy, 'https': proxy}), urllib.request.HTTPCookieProcessor(cookieJar))
 		else:
-			opener = urllib2.build_opener(urllib2.HTTPCookieProcessor(cookieJar))
+			opener = urllib.request.build_opener(urllib.request.HTTPCookieProcessor(cookieJar))
 		opener.addheaders = headers
 
 		try:
 			response = opener.open(url)
 			jsonResponse = response.read()
 		except:
-			print "Twitter weird response. Try to see on browser: https://twitter.com/search?q=%s&src=typd" % urllib.quote(urlGetData)
+			#print("Twitter weird response. Try to see on browser: ", url)
+			print("Twitter weird response. Try to see on browser: https://twitter.com/search?q=%s&src=typd" % urllib.parse.quote(urlGetData))
+			print("Unexpected error:", sys.exc_info()[0])
 			sys.exit()
 			return
 		
-		dataJson = json.loads(jsonResponse)
+		dataJson = json.loads(jsonResponse.decode())
 		
 		return dataJson
