+++ got/manager/__init__.py
@@ -1,2 +1,2 @@
-from TweetCriteria import TweetCriteria
-from TweetManager import TweetManager
+from .TweetCriteria import TweetCriteria
+from .TweetManager import TweetManager
