+++ got/models/__init__.py
@@ -1,1 +1,1 @@
-from Tweet import Tweet
+from .Tweet import Tweet
