--- got3/manager/__init__.py
@@ -1,2 +1,0 @@
-from .TweetCriteria import TweetCriteria
-from .TweetManager import TweetManager
