--- got3/models/Tweet.py
@@ -1,4 +1,0 @@
-class Tweet:
-	
-	def __init__(self):
-		pass
