--- got3/models/__init__.py
@@ -1,1 +1,0 @@
-from .Tweet import Tweet
